backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
ifcopenshell==0.8.2
google-generativeai==0.8.5
pydantic==2.11.3 